+++ tests/conftest.py
@@ -1,2 +1,4 @@
+# -*- coding: utf-8 -*-
+
 # Disable the sample tests for now
 collect_ignore = ['test_samples.py']